+++ hyracks-storage-am-lsm-invertedindex/src/main/java/edu/uci/ics/hyracks/storage/am/lsm/invertedindex/search/InvertedListPartitions.java
@@ -40,7 +40,7 @@ public class InvertedListPartitions {
     }
 
     @SuppressWarnings("unchecked")
-    public void reset(int numTokensLowerBound, int numTokensUpperBound) {
+    public void reset(short numTokensLowerBound, short numTokensUpperBound) {
         if (partitions == null) {
             int initialSize;
             if (numTokensUpperBound < 0) {
